requirements.txt
flask
twilio
python-dotenv
google-generativeai
requests
